move status functions to its own service
